+++ packages/categories/server/cats.js
@@ -1,0 +1,15 @@
+
+
+Meteor.publish('catsByName', function(name, type) {
+
+	if(this.userId)
+	{
+		var cur = K.findCatsByName(name, type);
+
+		console.log('Pub: catsByName', name, cur.count());
+
+		return cur;
+	}
+	else
+		this.ready();
+});