+++ packages/pois/client/Place_pois.js
@@ -15,7 +15,7 @@ Kepler.Place.include({
 		else
 			Meteor.subscribe('poisByPlace', self.id, function() {
 				
-				self.poisList = findPoisByLoc(self.loc).fetch();
+				self.poisList = K.findPoisByLoc(self.loc).fetch();
 
 				self._dep.changed();